move spyder helper
