+++ qcodes/utils/helpers.py
@@ -52,6 +52,7 @@ from .function_helpers import is_function
 from .json_utils import NumpyJSONEncoder
 from .path_helpers import QCODES_USER_PATH_ENV, get_qcodes_path, get_qcodes_user_path
 from .qt_helpers import foreground_qt_window
+from .spyder_utils import add_to_spyder_UMR_excludelist
 from .val_mapping import create_on_off_val_mapping
 
 log = logging.getLogger(__name__)
@@ -60,51 +61,6 @@ log = logging.getLogger(__name__)
 def warn_units(class_name: str, instance: object) -> None:
     logging.warning('`units` is deprecated for the `' + class_name +
                     '` class, use `unit` instead. ' + repr(instance))
-
-
-
-
-
-def add_to_spyder_UMR_excludelist(modulename: str) -> None:
-    """
-    Spyder tries to reload any user module. This does not work well for
-    qcodes because it overwrites Class variables. QCoDeS uses these to
-    store global attributes such as default station, monitor and list of
-    instruments. This "feature" can be disabled by the
-    gui. Unfortunately this cannot be disabled in a natural way
-    programmatically so in this hack we replace the global ``__umr__`` instance
-    with a new one containing the module we want to exclude. This will do
-    nothing if Spyder is not found.
-    TODO is there a better way to detect if we are in spyder?
-    """
-    if any('SPYDER' in name for name in os.environ):
-
-        sitecustomize_found = False
-        try:
-            from spyder.utils.site import sitecustomize
-        except ImportError:
-            pass
-        else:
-            sitecustomize_found = True
-        if sitecustomize_found is False:
-            try:
-                from spyder_kernels.customize import spydercustomize as sitecustomize
-
-            except ImportError:
-                pass
-            else:
-                sitecustomize_found = True
-
-        if sitecustomize_found is False:
-            return
-
-        excludednamelist = os.environ.get('SPY_UMR_NAMELIST',
-                                          '').split(',')
-        if modulename not in excludednamelist:
-            log.info(f"adding {modulename} to excluded modules")
-            excludednamelist.append(modulename)
-            sitecustomize.__umr__ = sitecustomize.UserModuleReloader(namelist=excludednamelist)
-            os.environ['SPY_UMR_NAMELIST'] = ','.join(excludednamelist)
 
 
 @contextmanager
